requirements.txt
freetype-py
pytest
matplotlib
pillow
